requirements.txt
argon2-cffi==20.1.0
async-generator==1.10
attrs==20.3.0
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.2.1
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cycler==0.10.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
fsspec==0.8.4
future==0.18.2
goto-statement==1.2
idna==2.10
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.16.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jsonlines==1.2.0
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kiwisolver==1.3.1
levenshtein==0.12.0
locket==0.2.0
lxml==4.6.2
MarkupSafe==1.1.1
matplotlib==3.3.3
mistune==0.8.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
notebook==6.1.5
numpy==1.19.4
oauthlib==3.1.0
packaging==20.7
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
partd==1.1.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pkg-resources==0.0.0
prometheus-client==0.9.0
prompt-toolkit==3.0.8
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.7.2
pyparsing==2.4.7
pyrsistent==0.17.3
PySocks==1.7.1
python-dateutil==2.8.1
python-twitter==3.5
pytz==2020.4
PyYAML==5.3.1
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
requests==2.25.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
six==1.15.0
snscrape==0.3.4
soupsieve==2.1
terminado==0.9.1
testpath==0.4.4
toolz==0.11.1
tornado==6.1
traitlets==4.3.3
tweet-preprocessor==0.6.0
urllib3==1.26.2
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
zipp==3.4.0
